synchronize with shell-skeleton

+++ script/skeleton/synchronize.sh
@@ -1,8 +1,12 @@
 #!/bin/sh -e
 
+DIRECTORY=$(dirname "${0}")
+SCRIPT_DIRECTORY=$(cd "${DIRECTORY}" || exit 1; pwd)
+# shellcheck source=/dev/null
+. "${SCRIPT_DIRECTORY}/../../lib/common.sh"
 TARGET="${1}"
 
-if [ "${TARGET}" = "" ]; then
+if [ "${TARGET}" = '' ]; then
     echo "Usage: ${0} TARGET"
 
     exit 1
@@ -14,9 +18,9 @@ if [ ! -d "${TARGET}" ]; then
     exit 1
 fi
 
-NAME=$(head -n 1 "${TARGET}"/README.md | awk '{ print $2 }' | grep --extended-regexp '^([A-Z]+[a-z0-9]*){2,}$') || NAME=""
+NAME=$(head -n 1 "${TARGET}/README.md" | awk '{ print $2 }' | grep --extended-regexp '^([A-Z]+[a-z0-9]*){1,}$') || NAME=''
 
-if [ "${NAME}" = "" ]; then
+if [ "${NAME}" = '' ]; then
     echo "Could not determine the project name."
 
     exit 1
@@ -33,26 +37,33 @@ else
 fi
 
 cp ./*.md "${TARGET}"
-cp ./*.py "${TARGET}"
 mkdir -p "${TARGET}/documentation"
 cp -R documentation/* "${TARGET}/documentation"
-mkdir -p "${TARGET}/debian"
-cp debian/* "${TARGET}/debian"
 mkdir -p "${TARGET}/script"
 cp -R script/* "${TARGET}/script"
-cp requirements.txt "${TARGET}"
+mkdir -p "${TARGET}/lib"
+cp lib/common.sh "${TARGET}/lib"
+cp .gitignore "${TARGET}"
 cp Vagrantfile "${TARGET}"
+cp ./*.py "${TARGET}"
+mkdir -p "${TARGET}/debian"
+cp debian/* "${TARGET}/debian"
+cp requirements.txt "${TARGET}"
 cp sonar-project.properties "${TARGET}"
 cp tox.ini "${TARGET}"
 cp .coveragerc "${TARGET}"
-cp .gitignore "${TARGET}"
 cp .pylintrc "${TARGET}"
 cp .pytest.ini "${TARGET}"
 cp .pytest-ci.ini "${TARGET}"
 cd "${TARGET}" || exit 1
-rm -rf script/skeleton
+echo "${NAME}" | grep --quiet 'Skeleton$' && IS_SKELETON=true || IS_SKELETON=false
+
+if [ "${IS_SKELETON}" = false ]; then
+    rm -rf script/skeleton
+fi
+
 DASH=$(echo "${NAME}" | ${SED} --regexp-extended 's/([A-Za-z0-9])([A-Z])/\1-\2/g' | tr '[:upper:]' '[:lower:]')
 INITIALS=$(echo "${NAME}" | ${SED} 's/\([A-Z]\)[a-z]*/\1/g' | tr '[:upper:]' '[:lower:]')
 UNDERSCORE=$(echo "${DASH}" | ${SED} --regexp-extended 's/-/_/g')
 # shellcheck disable=SC2016
-${FIND} . -type f -regextype posix-extended ! -regex '^.*/(build|tmp|\.git|\.idea|\.venv|\.tox|\.cache|\.vagrant|__pycache__|[a-z_]+\.egg-info)/.*$' -exec sh -c '${1} --in-place --expression "s/PythonSkeleton/${2}/g" --expression "s/python-skeleton/${3}/g" --expression "s/python_skeleton/${4}/g" --expression "s/pyskel/${5}/g" "${6}"' '_' "${SED}" "${NAME}" "${DASH}" "${UNDERSCORE}" "${INITIALS}" '{}' \;
+${FIND} . -regextype posix-extended -type f ! -regex "${EXCLUDE_FILTER}" -exec sh -c '${1} -i --expression "s/PythonSkeleton/${2}/g" --expression "s/python-skeleton/${3}/g" --expression "s/python_skeleton/${4}/g" --expression "s/pyskel/${5}/g" "${6}"' '_' "${SED}" "${NAME}" "${DASH}" "${UNDERSCORE}" "${INITIALS}" '{}' \;
